+++ backend/lib/reqBody.ts
@@ -1,0 +1,11 @@
+enum RequestType {
+    CREATE_ROOM,
+    JOIN_ROOM,
+    REJOIN_ROOM,
+}
+
+class RequestBody {
+    constructor(public reqType: RequestType) {}
+}
+
+export { RequestBody, RequestType }
